// src/app/user/models/user.ts
  export class Todo {
    title?: string;
    due_date?: string;
    priority?: string;
    status?: string;
  }

  export class Status {
    status?: string;
  }

  export class Transcript {
    status?: string;
  }
  
  export class User {
    name?: string;
    password?: string;
    email?: string;
  }